Updated `initdb.py` to use `init_connections(...)` from `dbmanager.py`
